fix: repo-wide review — an aliasing clone, stale claims, and shed duplicates
The repo-wide code review the working agreement asks for before a merge, run
across the whole fork-owned surface (JS hosts and drivers; platform and
webaudio C++; scripts; documentation), hunting broken, vestigial, redundant,
and anything leaning against a decision that has since closed.

BROKEN, fixed:

- webaudio Source::clone() retained and returned `this` — a "clone" that
  aliases, so stop() or setVolume() on the copy hit the original. Every other
  backend returns a real copy, and unlike this backend's other divergences it
  was undeclared. Now a copy constructor: same format, settings and static PCM,
  no host voice until it first plays, stopped — openal's semantics.
- touch-backend passed an int64 id to a %d format — varargs mismatch printing
  garbage for ids >= 2^31. Upstream's sdl/Touch.cpp has the same bug;
  fork-authored code can just be right: %lld with a cast.
- Two comment/code contradictions: fs-backend.h still declared directory
  enumeration "deliberately unimplemented" (real since fs_list landed), and
  Audio.h still declared end-of-buffer Sources unreclaimable (reapFinished()
  now reclaims static ones; the STREAM/QUEUE residue is the honest remainder).
- webgl-host's header claimed all strings answer from the real context
  (glGetString serves fixed identity strings) and counted 161 imports (162
  since glGetStringi); the game witness's failure message claimed
  "byte-identical" for a two-statistic comparison.

REDUNDANT, shed:

- wasi/graphics/driver.mjs and wasi/audio/driver.mjs duplicated
  wasi/platform/driver.mjs (whose header even records replacing four such
  copies) — and both carried the memory-detach hazard the shared one fixes:
  mem() evaluated before pump_in(), so a grow during the call wrote to a
  detached buffer. Deleting the copies deletes the bug. Callers retargeted.
- Three audio probes re-declared makeSineWav and resolvePlaywright, both
  exported by witness-harness.mjs for exactly this reason. Consolidated.
- w_draw_texture inlined the 2x2 four-texel texture that makeQuadTexture()
  builds; the helper's own comment said "shared". Now it is.

DRIFT, corrected — the biggest finding is one the review existed to catch: the
readme's front door still said the corpus witness was NOT in CI ("not wired in
yet", "needs a token scope"), written when that was true and stale the moment
the Human committed the workflow step. Now: CI-enforced, plainly. The same
sweep re-pointed every document that still called D4 open (DESIGN's decisions
table, EMBEDDING, fs-backend.cpp's #48 comment), updated the packaging rows to
the #7 recommended shape (one product file, host in the page) instead of
asserting single-.js as settled, fixed the handoff's 303/37-era numbers and
its "observed, not tested" evidence grade (the mapping is re-earned on every
push now), and marked the census table historical.

Scripts came back clean: all 55 verified, every referenced path resolves,
lua53-strlib linked in all fourteen LOVE-linking builds, EH gates uniform.

Left as-is, deliberately: recInt/recFloat duplicated across input-backend and
joystick-backend (two one-line static helpers; a shared header costs more than
it saves), and the two GL hosts' deliberate duplication — verified byte-level
in-sync across the whole shared surface including the new #51 block.

Evidence: all 20 graphics legs, audio (node + Chromium), input (node +
Chromium), and the corpus (306/34/15) re-run green against rebuilt artifacts;
a resample probe run end-to-end on the consolidated helpers.

+++ wasi/platform/touch-backend.cpp
@@ -56,7 +56,11 @@ const Touch::TouchInfo &Touch::getTouch(int64 id) const
 			return touch;
 	}
 
-	throw love::Exception("Invalid active touch ID: %d", id);
+	// id is int64; %d reads varargs as int, printing garbage for ids >= 2^31.
+	// Browser Touch.identifier values are small, but the format should not rely
+	// on that. (Upstream's sdl/Touch.cpp has the same mismatch; fork-authored
+	// code can just be right.)
+	throw love::Exception("Invalid active touch ID: %lld", (long long) id);
 }
 
 void Touch::onEvent(TouchEvent type, const TouchInfo &info)
